Patch for sheets preview window sometimes displaying improper verses
This fix should be considered temporary and isn’t really complete as it
essentially just skips the cache which is where the underlying problem
is happening. i.e. makePreview expects a ref that’s already parsed w/
context=0 and the cache returns the entire section.

That said — calling the cache in this function only seemed to only get
called when the ref matched matched the book title and not any of the
alternative titles, so this shouldn’t increase the number of calls to
the api in any significant way.