Fix for Creating Source Sheet input from Add To Sheet on Sidebar in Safari
user-select has inconsistent behavior between browsers in treatment of
input forms. This fix allows for the selected text to be added to a new
source sheet in Safari but the selected text appears to become
unhighlighted. The data is still sent, and I think this represents the
best scenario right now short of faking the highlighting using
background color triggered via `selectionchange` event in the Selection
API…