chore: elasticsearch responds best to user:pass@server format.
We found that specifying the user and pass with the basic_auth=("elastic", ELASTIC_PASSWORD) style argument didn't work correctly in 7.x.  It seems like it will work correctly in 8.x.  We currently can't go to 8.x because we depend of the elastic-dsl library, and that library is only current to 7.x.