feat(sheet): redirect (permanent, 301) the language of sheet - he for hebrew interface, and bi for english interface.
Implications 301 for SEO
Permanent change: 301 can be used for permanent URL changes. These include such things as site migrations or merging pages.
Link authority: a 301 redirect retains up to 90-99% of link authority. This is essential for SEO.
User experience: it is important to create a good transition for users following old bookmarks or links.